Successfully isolated problem with GUI (which almost derailed Jimbo Wales demo).  With noautohide="true" on the edit panel it is now impossible to use the edit dialog.
It needs to be determined why this is occuring, particularly because noautohide is considered important to faciliate looking up sources (in new windows) while editing refs.  noautohide was added back in revision 62 (i.e. diff -r 670ed22a9799 -r 550fff12e4ab) on Jan 22.  I know editing has not been broken all that time.   Further, I do not recall this happening when this noautohide was used briefly for debuggging before.

As a sidenote, having noautohide true causes a new corresponding Firefox window to appear when editing (or adding a new reference) on Xorg (at least on the Red Hat machine).  However, this is probably irrelevant here.

Note that noautohide does /not/ cause problems on the createnew dialog, which (along with the fact that breakage is recent) strongly indicates a JavaScript error.