Enhance preferences and menus.
Remove preference observer on unload (at least if prefs is not null).
Simplify a debug print statement.
Add a real preference dialog.  Currently the only preference is to include the edit summary
Add a right-click menu to the status bar element, with options for the preference window and toggling the sidebar.
Change minversion from 2.9 to 3.0
Minor tweaks to DTD